Improve detection whether SecureBoot is enabled
As the capacity of the kernel ring buffer is limited, it isn't very
robust to search for a specific string in the output of `dmesg(1)`. The
likelihood the determine the correct state of SecureBoot is extremely
reduced on machines with a high volume of kernel logging. Some examples
given, this can be either per configuration (`iptables(8)` LOG target),
broken hardware (kernel barking all the time because of broken ACPI
tables), or simply by high uptime of the machine.

It might be better to use `mokutil(1)` in this case which is provided
by the RPM of the same name. The existance of `mokutil(1)` can be
safely assumed on relevant machines as its RPM is a dependency of the
`shim-x64` RPM which in turn provides the UEFI bootloader.